refactor(launcher-cli): decompose docker runtime helpers
